# perf: replace goccy with gjson and store response items in a slice 