--- a/appender.go
+++ b/appender.go
@@ -269,7 +269,7 @@ func (a *Appender) flush(ctx context.Context, bulkIndexer *bulkIndexer) error {
 		return err
 	}
 	var docsFailed, docsIndexed, tooManyRequests, clientFailed, serverFailed int64
-	for _, info := range resp.Items {
+	for _, info := range resp {
 		if info.Error.Type != "" || info.Status > 201 {
 			docsFailed++
 			if info.Status >= 400 && info.Status < 500 {